+++ lib/smtp.sh
@@ -10,5 +10,4 @@ function check_smtp() {
   if [[ "${USE_SMTP}" == "TRUE" ]] && [[ -n "${ssmtp_cmd}" ]]; then
     sendmail_cmd="${ssmtp_cmd}"
   fi
-  trace "sendmail_cmd=${sendmail_cmd}"
 }